Extract getEnvironmentSetUp to it's own TestCase
