+++ tests/Integration/Database/EloquentLazyEagerLoadingTest.php
@@ -1,29 +1,15 @@
 <?php
 
-namespace Illuminate\Tests\Integration\Database\EloquentLazyEagerLoadingTest;
+namespace Illuminate\Tests\Integration\Database;
 
-use Orchestra\Testbench\TestCase;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Eloquent\Model;
 
 /**
  * @group integration
  */
-class EloquentLazyEagerLoadingTest extends TestCase
+class EloquentLazyEagerLoadingTest extends DatabaseTestCase
 {
-    protected function getEnvironmentSetUp($app)
-    {
-        $app['config']->set('app.debug', 'true');
-
-        $app['config']->set('database.default', 'testbench');
-
-        $app['config']->set('database.connections.testbench', [
-            'driver' => 'sqlite',
-            'database' => ':memory:',
-            'prefix' => '',
-        ]);
-    }
-
     public function setUp()
     {
         parent::setUp();
