+++ tests/Integration/Database/EloquentModelCustomEventsTest.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace Illuminate\Tests\Integration\Database\EloquentModelCustomEventsTest;
+namespace Illuminate\Tests\Integration\Database;
 
-use Orchestra\Testbench\TestCase;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Eloquent\Model;
@@ -10,21 +9,8 @@ use Illuminate\Database\Eloquent\Model;
 /**
  * @group integration
  */
-class EloquentModelCustomEventsTest extends TestCase
+class EloquentModelCustomEventsTest extends DatabaseTestCase
 {
-    protected function getEnvironmentSetUp($app)
-    {
-        $app['config']->set('app.debug', 'true');
-
-        $app['config']->set('database.default', 'testbench');
-
-        $app['config']->set('database.connections.testbench', [
-            'driver' => 'sqlite',
-            'database' => ':memory:',
-            'prefix' => '',
-        ]);
-    }
-
     public function setUp()
     {
         parent::setUp();
